+++ src/main/java/mchorse/aperture/client/gui/panels/modules/GuiAngleModule.java
@@ -28,28 +28,28 @@ public class GuiAngleModule extends GuiAbstractModule
 
         this.yaw = new GuiTrackpadElement(mc, (value) ->
         {
-            this.angle.yaw = value;
+            this.angle.yaw = value.floatValue();
             this.editor.updateProfile();
         });
         this.yaw.tooltip(IKey.lang("aperture.gui.panels.yaw"));
 
         this.pitch = new GuiTrackpadElement(mc, (value) ->
         {
-            this.angle.pitch = value;
+            this.angle.pitch = value.floatValue();
             this.editor.updateProfile();
         });
         this.pitch.tooltip(IKey.lang("aperture.gui.panels.pitch"));
 
         this.roll = new GuiTrackpadElement(mc, (value) ->
         {
-            this.angle.roll = value;
+            this.angle.roll = value.floatValue();
             this.editor.updateProfile();
         });
         this.roll.tooltip(IKey.lang("aperture.gui.panels.roll"));
 
         this.fov = new GuiTrackpadElement(mc, (value) ->
         {
-            this.angle.fov = value;
+            this.angle.fov = value.floatValue();
             this.editor.updateProfile();
         });
         this.fov.tooltip(IKey.lang("aperture.gui.panels.fov"));
